EES-5639 Moving the `CreateRelease` test out of `ReleaseVersionServicePermissionTests` into `ReleaseServicePermissionTests`
